+++ pkg/storegateway/bucket_store_metrics.go
@@ -58,9 +58,9 @@ type BucketStoreMetrics struct {
 	indexHeaderReaderMetrics *indexheader.ReaderPoolMetrics
 
 	// Quantify how much the projections optimization helps reduce labels sent to queriers.
-	originalLabelBytes  prometheus.Counter
-	reducedLabelBytes   prometheus.Counter
-	increasedLabelBytes prometheus.Counter
+	originalLabelBytes prometheus.Counter
+	reducedLabelBytes  prometheus.Counter
+	skippedLabelBytes  prometheus.Counter
 }
 
 func NewBucketStoreMetrics(reg prometheus.Registerer) *BucketStoreMetrics {
@@ -207,15 +207,15 @@ func NewBucketStoreMetrics(reg prometheus.Registerer) *BucketStoreMetrics {
 	})
 	m.originalLabelBytes = promauto.With(reg).NewCounter(prometheus.CounterOpts{
 		Name: "cortex_bucket_store_projection_original_label_bytes_total",
-		Help: "Total number of bytes of labels transferred to queriers.",
+		Help: "Total number of bytes of original labels transferred to queriers when projections are used.",
 	})
 	m.reducedLabelBytes = promauto.With(reg).NewCounter(prometheus.CounterOpts{
 		Name: "cortex_bucket_store_projection_reduced_label_bytes_total",
-		Help: "Total number of bytes of labels saved using projections.",
+		Help: "Total number of bytes of reduced labels transferred to queriers when projections are used.",
 	})
-	m.increasedLabelBytes = promauto.With(reg).NewCounter(prometheus.CounterOpts{
-		Name: "cortex_bucket_store_projection_increased_label_bytes_total",
-		Help: "Total number of bytes of labels increased using projections.",
+	m.skippedLabelBytes = promauto.With(reg).NewCounter(prometheus.CounterOpts{
+		Name: "cortex_bucket_store_projection_skipped_label_bytes_total",
+		Help: "Total number of bytes of labels transferred to queriers when projections are not used.",
 	})
 
 	return &m